+++ src/87source.js
@@ -14,11 +14,14 @@ yy.Source.prototype.toString = function() {
 }
 
 // SOURCE FILE
-yy.Source.prototype.execute = function (databaseid) {
+yy.Source.prototype.execute = function (databaseid,params,cb) {
 //	console.log(this.url);
-	loadFile(this.url, function(data){
+	loadFile(this.url, !!cb, function(data){
 //		console.log(data);
+		var res = 1;
 		alasql(data);
+		if(cb) res = cb(res);
+		return res;
 	}, function(err){
 		throw err;
 	});